+++ sample/07.GraphQL/gqlgen.yml
@@ -35,7 +35,7 @@ resolver:
 # gqlgen will search for any type names in the schema in these go packages
 # if they match it will use them, otherwise it will generate them.
 autobind:
-#  - "example/graphql/graph/model"
+ - "example/graphql/graph/model"
 
 # This section declares type mapping between the GraphQL and go type systems
 #
@@ -54,3 +54,7 @@ models:
       - github.com/99designs/gqlgen/graphql.Int
       - github.com/99designs/gqlgen/graphql.Int64
       - github.com/99designs/gqlgen/graphql.Int32
+  Todo:
+    fields:
+      user:
+        resolver: true
