+++ sample/07.GraphQL/graph/model/models_gen.go
@@ -7,13 +7,6 @@ type NewTodo struct {
 	UserID string `json:"userId"`
 }
 
-type Todo struct {
-	ID   string `json:"id"`
-	Text string `json:"text"`
-	Done bool   `json:"done"`
-	User *User  `json:"user"`
-}
-
 type User struct {
 	ID   string `json:"id"`
 	Name string `json:"name"`
