+++ sample/07.GraphQL/graph/model/todo.go
@@ -1,0 +1,8 @@
+package model
+
+type Todo struct {
+	ID   string `json:"id"`
+	Text string `json:"text"`
+	Done bool   `json:"done"`
+	User *User  `json:"user"`
+}
