+++ sample/07.GraphQL/graph/resolver.go
@@ -1,5 +1,7 @@
 package graph
 
+//go:generate go run github.com/99designs/gqlgen generate
+
 import "example/graphql/graph/model"
 
 // This file will not be regenerated automatically.
